# Conflicts:
#	sdk/src/main/java/com/horizen/account/state/EvmMessageProcessor.java
#	sdk/src/main/scala/com/horizen/account/api/http/AccountTransactionApiRoute.scala
#	sdk/src/main/scala/com/horizen/account/api/rpc/service/EthService.scala
#	sdk/src/main/scala/com/horizen/account/forger/AccountForgeMessageBuilder.scala
#	sdk/src/main/scala/com/horizen/account/state/AccountState.scala
#	sdk/src/main/scala/com/horizen/account/state/AccountStateReader.scala
#	sdk/src/main/scala/com/horizen/account/state/AccountStateView.scala
#	sdk/src/main/scala/com/horizen/account/state/StateTransition.scala

+++ sdk/src/main/scala/com/horizen/account/api/http/AccountTransactionApiRoute.scala
@@ -130,11 +130,11 @@ case class AccountTransactionApiRoute(override val settings: RESTApiSettings,
       applyOnNodeView { sidechainNodeView =>
         val valueInWei = ZenWeiConverter.convertZenniesToWei(body.value)
         val destAddress = body.to
-        // TODO actual gas implementation
+        // TODO add SuggestTipCap + BaseFee here from block
         val gasPrice = {
           val bestBlock = sidechainNodeView.getNodeHistory.getBestBlock
           if (bestBlock == null) BigInteger.ZERO
-          else BigInteger.valueOf(bestBlock.header.baseFee)
+          else bestBlock.header.baseFee
         }
         val gasLimit = GasUtil.TxGas
         // check if the fromAddress is either empty or it fits and the value is high enough
